Split api files and add more annotations
