+++ pkg/apis/fleet.cattle.io/v1alpha1/generic_map.go
