+++ pkg/apis/fleet.cattle.io/v1alpha1/groupversion_info.go
@@ -1,0 +1,6 @@
+// Copyright (c) 2021-2023 SUSE LLC
+
+// Package v1alpha1 contains API Schema definitions for the fleet.cattle.io v1alpha1 API group
+// +kubebuilder:object:generate=true
+// +groupName=fleet.cattle.io
+package v1alpha1
